ai resume analyzer/requirements.txt
fastapi
uvicorn
spacy
scikit-learn
pdfplumber
python-docx
pydantic
